requirements.txt
maafw==4.5.6
loguru
json-with-comments